renamed 'message.js' module to 'dialog.js' module, minor code changes
