+++ public/dialog.js
@@ -1,0 +1,25 @@
+const { dialog } = require('electron')
+
+let messageBoxOpen = false
+
+function showMessageBox(options, cb) {
+
+  if (messageBoxOpen) {
+    return
+  }
+
+  messageBoxOpen = true
+
+  dialog.showMessageBox(
+    options,
+    (selectedIndex) => {
+      try {
+        cb(selectedIndex)
+      }
+      finally {
+        messageBoxOpen = false
+      }
+    })
+}
+
+module.exports = { showMessageBox }
